+++ sickbeard/metadata/mede8er.py
@@ -129,9 +129,7 @@ class Mede8erMetadata(mediabrowser.MediaBrowserMetadata):
             logger.log(u"TVDB is down, can't use its data to make the NFO", logger.ERROR)
             raise
 
-        if not myShow:
-            logger.log(u'Show %s not found on %s ' % (show_obj.name, sickbeard.indexerApi(show_obj.indexer).name),
-                       logger.WARNING)
+        if not self._valid_show(myShow, show_obj):
             return
 
         # check for title and id
@@ -247,9 +245,7 @@ class Mede8erMetadata(mediabrowser.MediaBrowserMetadata):
             logger.log(u"Unable to connect to TVDB while creating meta files - skipping - " + ex(e), logger.ERROR)
             return False
 
-        if not myShow:
-            logger.log(u'Show %s not found on %s ' % (ep_obj.show.name, sickbeard.indexerApi(ep_obj.show.indexer).name),
-                       logger.WARNING)
+        if not self._valid_show(myShow, ep_obj.show):
             return
 
         rootNode = etree.Element("details")